Change everything to use new settings config method
The settings are now accessible through a module, the hub
and settings popup modify them; all other modules only read.
instead of transferring settings through signals, a settings
changed signal is emitted when they are changed and users
that need to check tha changes can read the settings without issues.

The commit includes general cleanup which this made feasible to do
that is too large scope to go into here