Do not use powershell script for updating the one dir mode
We do not have to use a powershell script as deleting the current directory
is not necessary, all of the contents can be moved out of it into
a temporary directory, and then the new release can be unpacked into the existing one.
This prevents any issues that arise from trying to delete the directory while the pyinstaller
exe process is alive.

This also help us get into a consistent state where the new process always cleans up
temp files